Rename OtrStore file in otrStore
